requirements.txt
Pillow~=9.3.0
cx_Freeze~=6.13.1